StyleGAN2_ada/requirements.txt
absl-py==0.12.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
filelock==3.0.12
gdown==3.13.0
google-auth==1.30.1
google-auth-oauthlib==0.4.4
grpcio==1.38.0
idna==2.10
importlib-metadata==4.5.0
joblib==1.0.1
Markdown==3.3.4
ninja==1.10.0.post2
numpy==1.20.3
oauthlib==3.1.1
opencv-python==4.5.2.52
Pillow==8.2.0
protobuf==3.17.2
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PySocks==1.7.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.1.0
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
tqdm==4.61.0
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
zipp==3.4.1
